The dataset is from Brady Brelinski, Founding Member of the Manhattan Chocolate Society. The data is also used in a Kaggle competition.
The dataset "chocolate" consists of 1500 samples about past chocolate ratings. Our problem is best framed as a 5-class, single-label alogorithm, which predicts whether a chocolate will receive a rating of one of five classes: 1 - unpleasant 2 - disappointing 3 - satisfactory 4 - premium 5 - elite We will predict the quality of a chocolate depending on the Company Name, Specific Bean Origin, REF, Review Year, Cocoa Percentage and Bean Type. So, it is a supervised learning model, in which we use a classification algorithm to learn from a labeled dataset, providing answer key that the alogrithm can use to evaluate its accuracy on training data.